Re-add DateRange filtering for Gcal
Craig argued that the DateRange filtering should be done in BAMRU.net
and not in this conversion code.  Indeed I agree with this assessment.
BAMRU.net csv output should accept a start/finish range in the url, like
http://bamru.net/public/calendar.csv?start=2015-02-01&finish=2015-06-06.

But when I looked at the BAMRU.net code, I was sad to find it always
outputs the entire calendar, and does not allow any date-range
filtering.

The current BAMRU.net system is at end-of-life, and I don't want to make
any more changes to this code.  So for now, date-range filtering must be
done in this conversion code.

IMHO filtering the Gcal date-range is important for two reasons:

1) For fast rebuilding: The unfiltered list has 450+ events, while the
filtered list has ~45 events.  The filtered list rebuilds 10X faster,
which is important for development/testing/debugging.  The downside is
less Gcal data, but I've never had anyone complain about that.

2) To keep the Gcal data in sync with the HTML calendar data at
http://bamru.org/calendar.

For the future, there are two simplifications that can be made:

1) Move the DateRange filtering from the 'refine' task to the
bnet/download task.  This will isolate DateRange filtering to a single
place in the code.

2) BAMRU.net V2 supports date range filtering, and even better supports
relative date-range filtering.  For example:

http://v2.bamru.net/public/calendar.csv?start=now-2months&finish=now+14months

With this change, we will be able to remove all DateRange filtering
fromt this code base.