Fix cron bug
In 'production' mode, Cron is supposed to exit if the BNET CSV data is
unchanged from the last run.  There was a bug that caused Cron to run
end-to-end even if the data had not changed.

The following changes were made:
- Rakefile: add a timestamp to the log, for easier reading
- download: fix an old_text/new_text comparison bug
- cron    : force the use of MM_ENV instead of RAILS_ENV

Cron documentation is on GitHub at the javan/whenever repo.